Automatically update org agenda files
Since we are using Denote as our source for note handling, it makes
sense to use these files for org agenda as well. This change adds all
Denote notes to org agenda on startup if they contain the "todo"
tag. Notes are also automatically added and removed when files are
added or renamed on disk.